Setup snapshot tests
I setup only the snapshot tests for now, and I reorganized the tests
into various categories. The idea is that I want to complete the
structure and styling of the home and shopping pages so that I can then
create the snapshots. Then, I can create the route test so that I can
work on the routing of the pages. Finally, I will create the logic tests
so that I can work on the logic of the shopping cart.