Create shopping page
I created the shopping page and made it responsive across all devices.
Additionally, I created the product component and started writing the
logic for the product props. Finally, I added functions for both the
shopping and home components that change the background colour of the
html, body, and root id. The home page requires a black background,
whereas the shopping page requires a white background.