Build side bar
The side bar/shopping cart is built, and ready for deployment. The side
bar is meant to extend from the left side of the screen over the rest of
the page, and a dark overlay that is slightly transparent is placed
under the side bar and above the page. The shopping cart when empty,
will display a message contained within the class
"shopping-cart-empty-message". When the shopping cart is not empty, a
container represented by class "side-bar-main-content" will be visible,
and will contain the chosen products, represented by the new
ShoppingCartProduct components. As I build the logic for this
application, the subtotal will change based on which products are chosen
by the user, as well as the quantity.