Fix existing tests
The main issue with my routing tests was that the root id element was
not being found in the DOM with regards to the blackPageBackground() and
whitePageBackground() functions in the Home and Shopping components. I
therefor got rid of the lines of code meant to style the root id element
because it wasn't even needed as I was already styling the html and body
elements with these functions. I was also having issues with the first
Home component snapshot test. It turns out that without providing a sort
of routing context to the Jest snapshot test, there could arise some
issues. Because I was already testing the initial routing of the Home
component upon startup of the application (using a snap shot in
tandem with it), I decided to remove the initial snapshot test, as it
was made redundant and was not even working.