Set up logic tests
I set up all the logic tests that I deemed necessary for future
undertaking of working out the logic. The two main testing scenarios I
set up were; 1. adding a product to the shopping cart 2. Changing the
product quantity and testing for the total price, and testing for the
removal of a shopping cart item based on the click event of the cancel
button, and the decrement button to 0. These tests mock the shopping
component, simulating the clicking event of the various buttons, and
make assertions on certain values found in the dom.