Make side bar interactive and header sticky
I updated the shopping css, and created some onClick functions in the
Shopping component, so that the sidebar can be interacted with (retract
and extend). Additionally, the header of the shopping component has been
made sticky, so that the user has access to those buttons no matter
where they're at in the page.