Add logic to add product to state and UI cart
I added state object and variables, a logic function, and a useEffect
hook which update shopping cart state of the user. I also made use of a
style function called addProductsToPhysicalCart which adds the product
added by the user to a cartItems state array. This array is mapped over
after every re-render and ShoppingCartProducts instances are created
and rendered into the shopping cart of the side bar.