Add logic to remove product from state and UI cart
Added a new state variable "productToBeRemoved", and a remove product
function, which is sent as a prop to the ShoppingCartProduct component.
When the x button is clicked on the shopping cart product, this function
is called which updates the productToBeRemoved state variable, which now
triggers the useEffect hook. I updated the useEffect hook to remove the
product from the shopping cart UI and from state if the
productToBeRemoved state variable does contain something.