Add quantity, and quantity selection functionality
Now the quantities for the selected products are shown in the shopping
cart. The quantities can also be incremented and decremented. Finally,
if the quantity is decremented below 1, the product is removed from the
shopping cart. New call back functions were created and sent as props
to the ShoppingCartProduct component, which are triggered on click.
These functions set new state variables, which trigger the useEffect
hook, which has been modified to make changes to the
cartProductQuantities state object.