Add cart toggle modes functionality
Now when the user tries to view the cart without selecting any products,
the cart shows a cart empty message. As soon as an item is added to the
cart, the message is gone and the cart gets populated with the
selection. Finally, I made it possible to retract the side bar by
clicking on the overlay (the part of the web page outside the side bar).