Add functionality for total number of items
I added a state variable that keeps track of the total number of items
in the shopping cart. This state variable is updated after every
re-render in the useEffect hook. The total number of shopping cart items
is displayed next to the shopping cart logos of the shopping page header
and the side bar header.