Add query parameters in image urls
Because some of my images are not rendering in github-pages, I added
this query parameter:

?v=1

at the end of each problematic image url. This is done because github
pages probably cached an older version of my project, and the query
parameter at the end of a url essentially forces github pages to fetch
the updated images.